Differentiate between payee and narration
Previously, the payee was not individually reported and the narration
was a combination of the transaction type and the payee. The transaction
purpose given in the input file was ignored.

Now, the payee is reported as such and the narration is a combination of
the transaction type and purpose.

For example:

Before:

2018-01-01 * "Lastschrift REWE Filialen Voll"

After:

2018-01-01 * "REWE Filialen Voll" "Lastschrift REWE SAGT DANKE."